✨ feat(Dashboard): refactor dashboard structure by creating a new products page and simplifying the main dashboard layout for improved navigation
